pyproject.toml
[tool.poetry]
name = "szukajwarchiwach"
version = "0.1.0"
description = ""
authors = ["Jan Szumiec <jan.szumiec@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pillow = "^7.1.2"

[tool.poetry.dev-dependencies]
requests = "^2.23.0"
beautifulsoup4 = "^4.8.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
